requirements.txt
# Core ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
lightgbm>=4.0.0

# Feature engineering
scipy>=1.11.0

# Explainability
shap>=0.43.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Data handling
pyarrow>=14.0.0
requests>=2.31.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.13.0

# Development
pytest>=7.4.0
tqdm>=4.66.0
rich>=13.7.0
pyyaml>=6.0.0
xgboost>=2.0.0
catboost>=1.2.0
requests>=2.31.0
